requirements.txt
asgiref==3.11.0
Django==5.2.8
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.1
psycopg-binary==3.3.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.4
timedelta==2020.12.3
